# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Panier API Integration",
        "description": "API endpoints for communicating with the Panier API - ZIMRA Fiscalisation Support.\r\n\r\n## Rate Limit\r\n500 API calls every 5 minutes for each unique IP Address.\r\n\r\n## Authentication\r\nRequires APP-ID and API-KEY headers obtained from your Panier company account's Developer API section.\r\n\r\n## ZIMRA Fiscalisation\r\nThis API supports ZIMRA Fiscal Invoices, Credit Notes and Debit Notes based on ZIMRA Fiscal Device Gateway API Specs v7.2.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/currency/create": {
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Create currencies",
                "description": "Supports 119 currencies",
                "operationId": "3befcae24f3adab432f139d0fe196747",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "code"
                                            ],
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "USD"
                                                },
                                                "exchange_rate": {
                                                    "type": "number",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/currency/update": {
            "put": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Update currencies",
                "operationId": "8c336449137c43fff29dc9978c6e933a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "exchange_rate": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/currency/search": {
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Search currencies",
                "operationId": "e9e4147969fda2609635d41e68eb89c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "query": {
                                                "type": "string",
                                                "default": "*"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/currency/delete": {
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Delete currencies",
                "operationId": "f61ff6a3adf368b7ca33add1020a9ca7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "AppId": {
                "type": "apiKey",
                "description": "Panier APP-ID from Developer API section",
                "name": "APP-ID",
                "in": "header"
            },
            "ApiKey": {
                "type": "apiKey",
                "description": "Panier API-KEY from Developer API section",
                "name": "API-KEY",
                "in": "header"
            }
        }
    },
...
}